Implement Storage.java to store TaskList and Driver.java to process commands
