+++ src/main/java/tasks/Deadline.java
@@ -1,6 +1,7 @@
 package tasks;
 
 public class Deadline extends Task {
+    private static final long serialVersionUID = 8193173341399324817L;
     String by;
 
     public Deadline(String description, String by) {
