+++ src/main/java/tasks/ToDo.java
@@ -1,6 +1,8 @@
 package tasks;
 
 public class ToDo extends Task {
+    private static final long serialVersionUID = -8037857826265085451L;
+
     public ToDo(String description) {
         super(description);
     }
